Add Purge button as its own row in the supermod Moderator Actions sidebar
